+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.0.25
 
 - Elaborate, Meta-Sample
+- Hello, Spinbox!
+- Support `spinbox` `command {}` and `on('increment') {}` / `on('decrement') {}`
 - Make `grid sticky: 'nsew', column_weight: 1` the default for all widgets (column weight is only applied to the first widget in a series of siblings)
 - Support `text` widget with `text` attribute/data-binding just like `entry`
 - Set `background '#ececec' if OS.mac?` on `root` by default